BUGFIX - adjust multiclass nonspecific function to Tabular
As a new feature, the tabular model adapter is using the function _adjust_multiclass_nonspecific to adapt multiclass models. However, since this function was initially designed to Image CF generator, the factual and prediction function (mic) is different. This commit fix the input from Tabular data to be compatible with the input required by the function.