Modify change factor calculation to favor large distances changes
Since it's more benefitial to allow big changes that cost low positive or high negative prediction changes, this commit makes this modification. Where small positive changes that greatly change the distance are favored and large negative changes that largely change the distance are also favored. The change in distance must be interpreted as how much closer the CF is from a factual, therefore, larger changes are better.