New random approach
This commit creates a subtype of the Random method. It is a true random since it does not increment changes over iteration, each iteration is a totally new random CF candidate from the received factual. Since this method is not biased (as the latter). The old random method is now called random-sequential, since the changes are done randomly but sequentially.